adding the a guard clause to check if there is a similar request pending then cancel the subsequent one until the pending on is resolved...

+++ CLIENT/pages/tenants/resources/js/classes/tenantsClasses.js
@@ -519,6 +519,8 @@ class UserInterface extends UserinterfaceUtilities {
       "deletingTenant"
     );
 
+    if (!openLoader) return;
+
     const requestOptions = {
       mode: "cors",
       method: "DELETE",
